dev release 1.0.1

+++ libft/src/ft_string/ft_strcat.c
@@ -6,7 +6,7 @@
 /*   By: llornel <llornel@student.21-school.ru>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/15 13:46:24 by llornel           #+#    #+#             */
-/*   Updated: 2022/02/18 12:12:11 by llornel          ###   ########.fr       */
+/*   Updated: 2022/02/19 21:22:08 by llornel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strcat(char *dest, const char *src)
 {
-	size_t	i;
-	size_t	j;
+	int	i;
+	int	j;
 
 	i = 0;
 	while (dest[i])
