+++ libft/src/ft_string/ft_strcmp.c
@@ -6,7 +6,7 @@
 /*   By: llornel <llornel@student.21-school.ru>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/15 13:47:39 by llornel           #+#    #+#             */
-/*   Updated: 2022/02/18 12:12:16 by llornel          ###   ########.fr       */
+/*   Updated: 2022/02/19 21:21:49 by llornel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strcmp(const char *s1, const char *s2)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s1[i] && s2[i] && (s1[i] == s2[i]))
